# [single-shot] benchmarking the latency @open sesame 09/09 17:11