Added benchmarking the latency and running times in the test code
This compares old vs new singleshot api - showing reduction in latency and running-time

These values averaged over 10 continuous runs

|  | Open (us)      | Invoke (ms)           | Close (us)  |
| --- |:-------------:|:-------------:|:-----:|
| New (cache warmup) |  658   | 5195 | 206 |
| Old (cache warmup) |  1228 | 5199   | 5245 |
| New (no warmup) | 1653 | 5205  | 201 |
| Old (no warmup) | 7201 | 5225  | 5299  |


These values are just for the first run

| | Open (us)      | Invoke (ms)           | Close (us)  |
| --- |:-------------:|:-------------:|:-----:|
| New  |  12326   | 5231 | 2347 |
| Old |  58772 | 5250   | 52611 |

Working on #1547

Self evaluation:
Build test: [x]Passed [ ]Failed [ ]Skipped
Run test: [x]Passed [ ]Failed [ ]Skipped

Signed-off-by: Parichay Kapoor pk.kapoor@samsung.com