@kparichay 2019-09-05
> You may add the experimental results at the README.md in capi src directory

I will update this PR to add the results in README.md