@kparichay 2019-09-05
V2:
Code Cleaning
Using `g_get_real_time` instead of `chrono`