@kparichay 2019-09-05
V3:
Added README.md for `api/capi/src`

V2:
Code Cleaning
Using `g_get_real_time` instead of `chrono`